+++ CHANGELOG.md
@@ -46,6 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Completed task `phase-4` (engine)
 - Completed 3 tasks (engine)
 - Completed 4 tasks (engine)
+- Completed task `phase-8` (engine)
 
 ### Added